// src/app/checklist.interface.ts
export interface Checklist {
  orden_de_servicio: string;
  empleado: string;
  lectura_codigos_foto?: string;
  servofreno_foto?: string;
  pedal_freno_foto?: string;
  pedal_estacionamiento_foto?: string;
  cinturon_seguridad_foto?: string;
  cuadro_instrumentos_foto?: string;
  aire_acondicionado_foto?: string;
  bocina_claxon_foto?: string;
  iluminacion_interior_foto?: string;
  iluminacion_externa_foto?: string;
  limpiaparabrisas_foto?: string;
  limpia_medallon_foto?: string;
  neumaticos_friccion_foto?: string;
  otros_vehiculo_en_piso_foto?: string;
  estado_fugas_aceite_foto?: string;
  estado_nivel_calidad_lubricante_transmision_foto?: string;
  estado_nivel_calidad_lubricante_diferencial_foto?: string;
  cubrepolvos_flechas_foto?: string;
  componentes_direccion_foto?: string;
  componentes_suspesion_foto?: string;
  sistema_escape_completo_foto?: string;
  sistema_alimentacion_combustible_foto?: string;
  filtro_combustible_foto?: string;
  control_fugas_direccion_hidraulica_foto?: string;
  otros_altura_total_foto?: string;
  rodamiento_mazas_rueda_foto?: string;
  holgura_partes_suspension_rueda_foto?: string;
  control_neumaticos_desgaste_presion_foto?: string;
  profundidad_foto?: string;
  presion_foto?: string;
  otros_altura_media_foto?: string;
  nivel_calidad_aceite_motor_foto?: string;
  filtro_aire_foto?: string;
  filtro_polen_foto?: string;
  filtro_pcv_foto?: string;
  valvula_pcv_foto?: string;
  bujias_encendido_foto?: string;
  cables_bujias_bobinas_ignicion_foto?: string;
  nivel_anticongenlante_foto?: string;
  tapon_radiador_foto?: string;
  mangueras_sistema_foto?: string;
  Id_empleado?:number;
  IdVehiculo?:number,
  Id_ordendeservicio:number;
  numeroserie: string;
  Id:number;
  Nombre:string;
  No_serie:string;
  id_ordendeservicio:number;
  fecha:Date;
}